+++ src/main/java/blackjack/domain/participant/player/AddCardOrNot.java
@@ -1,4 +1,4 @@
-package blackjack.controller;
+package blackjack.domain.participant.player;
 
 public enum AddCardOrNot {
     YES,
