+++ src/main/java/blackjack/domain/participant/player/CardDecisionStrategy.java
@@ -1,6 +1,5 @@
 package blackjack.domain.participant.player;
 
-import blackjack.controller.AddCardOrNot;
 import blackjack.domain.participant.Participant;
 
 @FunctionalInterface
